Add date to rating
