+++ src/routes/browse.js
@@ -9,6 +9,7 @@ import MultiStepFormPage from "../components/multistepformpage";
 import { useAuth } from "../contexts/authContext";
 import { useNavigate } from "react-router-dom";
 import { getThumbnail, getColors, getMaterials } from "../backend";
+import { getDate } from "../utils";
 
 export default function Browse() {
   const [availableColors, setAvailableColors] = useState([]);
@@ -109,14 +110,6 @@ export default function Browse() {
 
   const isFilterSelected = (category, label) => {
     return filters[category].includes(label);
-  }
-  const getDate = () => {
-    const date = new Date();
-    const day = String(date.getDate()).padStart(2, '0');
-    const month = String(date.getMonth() + 1).padStart(2, '0'); // Months are zero-based
-    const year = date.getFullYear();
-
-    return `${day}/${month}/${year}`;
   }
 
   const onSubmit = () => {
